Add version bump commands to cli
Added ability to bump version of iku. New version is written to
package.json. Major, minor, and patch version can be bumped. When
bumping a minor or major version the subversions below it will be
reset to zero.

Minor versions should never exceed 9. We will just keep adding
patches until we are ready to bump the major. So we should never
see 0.10.4 however 0.9.233 is acceptable. Then when ready we will
bump to 1.0.0

Here are some examples:
  - 0.0.1   -> 0.0.2
  - 0.2.13  -> 0.3.0
  - 0.9.4  -> 1.0.0